+++ lib/ts/recipe/thirdparty/components/features/signInAndUpCallback/index.tsx
@@ -93,7 +93,7 @@ const SignInAndUpCallback: React.FC<PropType> = (props) => {
                     .validateGlobalClaimsAndHandleSuccessRedirection(
                         {
                             action: "SUCCESS",
-                            isNewPrimaryUser: response.createdNewRecipeUser && response.user.loginMethods.length === 1,
+                            createdNewUser: response.createdNewRecipeUser && response.user.loginMethods.length === 1,
                             isNewRecipeUser: response.createdNewRecipeUser,
                             newSessionCreated:
                                 session.loading ||