+++ test/unit/recipe/thirdparty/signInUp.test.tsx
@@ -88,7 +88,7 @@ describe("ThirdParty.SignInAndUp", () => {
                     successRedirectContext: {
                         action: "SUCCESS",
                         isNewRecipeUser: false,
-                        isNewPrimaryUser: false,
+                        createdNewUser: false,
                         redirectToPath: undefined,
                     },
                 },