+++ test/unit/recipe/thirdpartyemailpassword/signInUp.test.tsx
@@ -76,7 +76,7 @@ describe("ThirdPartyEmailPassword.SignInAndUp", () => {
                     successRedirectContext: {
                         action: "SUCCESS",
                         isNewRecipeUser: false,
-                        isNewPrimaryUser: false,
+                        createdNewUser: false,
                         redirectToPath: undefined,
                     },
                 },