+++ test/unit/recipe/thirdpartypasswordless/signInUp.test.tsx
@@ -77,7 +77,7 @@ describe("ThirdPartyPasswordless.SignInAndUp", () => {
                     successRedirectContext: {
                         action: "SUCCESS",
                         isNewRecipeUser: false,
-                        isNewPrimaryUser: false,
+                        createdNewUser: false,
                         redirectToPath: undefined,
                     },
                 },